Adapted the behaviour of output and input to exactly mirror the old snowpack operational behaviour, writing to and from the same location or reading from one writing to another is now possible - please set the corresponding paths in the input and output section of the io.ini:
[INPUT]
SNOPATH = current_snow

[OUTPUT]
SNOPATH = current_snow

In operational mode an aggregated profile for the flatfield is to be written into a file for subsequent upload to the DB by the SqlLoader (Oracle). You can activate this feature, by specifying the plugins you want to use for writing profiles:
[OUTPUT]
PROFILE = IMIS             #profile will only be written to files that are to be uploaded to the DB
PROFILE = ASCII           #profile will only be written to default ASCII files (the default behaviour if nothing is specified)
PROFILE = ASCII IMIS   #profile written in both formats!