The SnowpackConfig object now differentiates between a [Snowpack], [SnowpackAdvanced], [Input] and [Output] section: All ini keys related to Snowpack used to be read from the [Paramters] section. Now the keys are grouped into four different sections, making them easier to manage.
All modules in Snowpack reading the SnowpackConfig object needed to be adapted in order to read their local variables from the correct sections in the ini file.

FIXED_HN_DENSITY is now a parameter to be configured by the user in the [SnowpackAdvanced] section.

NOTE: snowpack-opera needs to be updated to the newest version.