cherry-picking commit 5e2884b
Added two new keys to control the inflate/deflate feature

With SNOWPACK::INFLATE_ALLOW it is possible to turn on/off inflate/deflate both
in operational and research mode (note that by default, it is turned off in
research and on in operational). If SNOWPACK::INFLATE_INFO is set to true, then
a message is printed every time inflate/deflate changes the mass balance.

Otherwise, the OSWR fields in the exmaple smet files have been renamed as RSWR.