Using computed ISWR in the ILWR parametrization
In late 2021, the cloudiness computation in MeteoIO was improved in
order to better handle the interpolation over the nights (since it is
based on the measured ISWR being compared with the potential ISWR, this
can not be computed at night). This lead to strong regressions for many
simulations (as reported by Fabiano Monti). In early 2022, a fix was
committed (MeteoIO dc347ac75e528b1ed1c1fdfea4841d4afdaefaec): the
terrain shading had to be taken into account in order to prevent
interpolating a cloudiness computed while in the shade of a neighbouring
mountain. This shading can be provided either as a horizon file or
through a DEM (and the computed horizon can then be saved in a horizon
file for quicker reuse later on).

Unfortunatelly, this was not enough for the IMIS stations that don't
measure ISWR but RSWR instead (although there are less impacted since
for most of the season SNowpack will rely on their measured TSS as
boundary condition). This problem had actually been hidden away be using
the inflate/deflate mechanism in the operational simulations (so
although all melt phases tended to be very bad, it was not visible).

This commit addresses this situation: now the ISWR that is computed in
Snowpack (based on the parametrized albedo) is use to call once more
potential ILWR data generators in MeteoIO in order to parametrize the
cloudiness and the ILWR. This is mostly transparent to the user but one
has to keep in mind that an ILWR data generator must be declared that
relies on ISWR even if no ISWR is measured...

This has been validated at WFJ and LARET (where both iswr and ilwr are
measured) as well as MUT2, TUM2, TUJ2, SCA2, DIA2, CHA2, VDS2, CON2 over
their respective whole available data period and the results are very
good (often around 1 day of offset between the measured HS and the
modeled HS, without using any inflate/deflate). The validation also
included LUM2 where some years are quite bad due to other issues
(offsets on TSS, possible avalanches triggering around the station,
etc).