Improving the coupled phase changes code.
First, all variables dth_i_up and dth_i_down now relate to half the element. Second, not lumping both together, but keep them attached to the node. When limiting the amount of melt given the amount of ICE and WATER, give priority to the phase changes at the upper element, and take the remaineder to limit the lower element.